+++ src/modules/dashboard/components/Dialogs/StakeExpenditureDialog/StakeExpenditureDialog.css.d.ts
@@ -1,6 +1,5 @@
 export const submitButtonWidth: string;
 export const padding: string;
-export const dialogContainer: string;
 export const title: string;
 export const forceContainer: string;
 export const stakeContainer: string;
