+++ src/modules/dashboard/components/Dialogs/StakeExpenditureDialog/index.ts
@@ -1,0 +1,1 @@
+export { default } from './StakeExpenditureDialog';
